added data table actions
