+++ src/components/ui/tables/data-table/columns.tsx
@@ -13,6 +13,7 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { MoreHorizontal } from "lucide-react";
 import { useRouter } from "next/router";
+import { TableActions } from "./table-actions";
 
 export const processColumns: ColumnDef<ProcessType>[] = [
   {
@@ -31,41 +32,46 @@ export const processColumns: ColumnDef<ProcessType>[] = [
     accessorKey: "assignee",
     header: "Assignee",
   },
-  {
-    accessorKey: "assignee",
-    header: "Created By",
-  },
+  // {
+  //   accessorKey: "assignee",
+  //   header: "Created By",
+  // },
   {
     id: "actions",
     cell: ({ row }) => {
       const process = row.original;
+      const { id } = process;
       //   const router = useRouter();
+      if (!process.id) {
+        console.error("Missing process ID", process);
+        return null;
+      }
+      return <TableActions processId={id} />;
+      // const handleEditClick = (processId: string) => {
+      //   // Navigate to the /dashboard/update-tasks route with process.id as a query param
+      //   console.log("clicked", processId);
+      //   // router.push(`/dashboard/update-task?processId=${processId}`);
+      // };
 
-      const handleEditClick = (processId: string) => {
-        // Navigate to the /dashboard/update-tasks route with process.id as a query param
-        console.log("clicked", processId);
-        // router.push(`/dashboard/update-task?processId=${processId}`);
-      };
-
-      return (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="ghost" className="h-8 w-8 p-0">
-              <span className="sr-only">Open menu</span>
-              <MoreHorizontal className="h-4 w-4" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <DropdownMenuLabel>Actions</DropdownMenuLabel>
-            <DropdownMenuItem onClick={() => handleEditClick(process.id)}>
-              Update Task
-            </DropdownMenuItem>
-            {/* <DropdownMenuSeparator /> */}
-            {/* <DropdownMenuItem>View customer</DropdownMenuItem>
-            <DropdownMenuItem>View payment details</DropdownMenuItem> */}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      );
+      // return (
+      //   <DropdownMenu>
+      //     <DropdownMenuTrigger asChild>
+      //       <Button variant="ghost" className="h-8 w-8 p-0">
+      //         <span className="sr-only">Open menu</span>
+      //         <MoreHorizontal className="h-4 w-4" />
+      //       </Button>
+      //     </DropdownMenuTrigger>
+      //     <DropdownMenuContent align="end">
+      //       <DropdownMenuLabel>Actions</DropdownMenuLabel>
+      //       <DropdownMenuItem onClick={() => handleEditClick(process.id)}>
+      //         Update Task
+      //       </DropdownMenuItem>
+      //       {/* <DropdownMenuSeparator /> */}
+      //       {/* <DropdownMenuItem>View customer</DropdownMenuItem>
+      //       <DropdownMenuItem>View payment details</DropdownMenuItem> */}
+      //     </DropdownMenuContent>
+      //   </DropdownMenu>
+      // );
     },
   },
 ];
